Hotel Management Demo Application
MVC application, bring to mind all the core functions. I created new controller which name is HotelManagementController, that is empty. Then I created new folder in Views - HotelManagement and then in the folder I created Index.cshtml file. I describe some html text. Then I created new model with name Manage. In the index method, of the controller, I created new object of Manage and use all the properties I added in the class. At the end I return the view with the name of the manage object. In Views/Index.cshtml, I added the using that I need to use the models . With @ symbol I said to the coder,that I will use c# code and display the properties with @model. property name. The relation is - the controller is trying to find View folder in the VIEWS with the same name of the controller name. In the folder of the view is trying to find Index.cshtml, by default to display the View.