Create method Added
Using bootstrap classes, I created nice looking form. The method is post, and the action is CreatePerson. The controller I'm using is HotelManagement. In the HotelManagement Controller, I created new static list of persons, then return the persons in Index view. I created "create" View to see the form.
Using data binding with "asp-for" to get the data from the input's field.
CreatePerson method is for creating the persons. It adds "manage" to the list.
When you click the save button, you are return to the index view to see the persons. Making the persons name, age and idnum to be seen, I wrote the new "using" List<> and use foreach loop, looping through the Model.
Then in html <li>, I print the data.