Relationships in ASP,Net7 MVC
I created 4 models. For one to one relationship, I used Character and Backpack models. In Character I implemented properties - Id,Name, Backpack of type Backpack.
In Backpack the properties are - Id, Description, CharactedId and Character of type Character. That's the relationship one to one.

For one to many, I used Character and Weapon models. In character I added list of Weapon weapons. In Weapon model, I added properties - Id, Name, CharacterId and Character of type Character. That's the relationship one to many.

For many to many, I used Character and Faction models. In Character, I added list of Faction Factions. In Faction I added properties - Id,Name and list of Characters. That's the relationship many to many.

Then I created separated folder Data , in which I added new ApplicationDbContext and I write the connection between the app and the sql in ApplicationDbContext. Then I write the DbSets<> with the models. I uploaded the migrations files too,to see the tables in.