Adding Views
In the ClubController, I get the data from the database, using constructor with parameter applicationDbContext.
In Index Method I get the clubs data, using the context I get from the applicationDbContext and make it ToList. I return the clubs and now the View part. Firstly I write the using which is of type Ienumerable<Club>. Secondly ,I use BootStrap classes for the design and one Foreach loop, to print all the cards. Using the @ symbol to print all the props I need. For races is absolutely the same.
Adding two navigation links to clubs and races and removing the one to privacy. For Club is club controller with action Index and for Race is race controller with action Index.