Detail Page.
I make a good design, thanks to bootstrap library. I created new Detail method with parameter Id. I got the Club only not "List" of clubs. Using .Include()EF function to get the Address. and the other function FirstOrdefault and make the clubId equal to the parameter Id.
At the end I return the club. Absolutely the same method is for Race, only the name is changed. As I said with the help of the bootstrap library, I created a good design of the View Details. I import the using - Club and with Foreach loop, I print all the information needed.  With the help of the tag helpers and anchor tag, I made the view button reactive using the Controller Name for clubs - Club , The Action name for clubs - Detail, and the asp-route.id with the @symbol and foreach to take the selected Id.