Dependency Injection & Interface & Repository added
Firstly, I created Interface folder and Repository folder. Repository is the almost the same like Services, but Repository is named when the application exchanges data with database.
Secondly, I created Interface and create 3 methods and CRUD operations. The same in the RaceRepository. Then, I created new class ClubRepositry and inherit from IClubRepositry "from the interface repository". I Created constructor in which, I assign the Database with name _context. Using the database, I wrote all the CRUD methods. Now...
For Getting all clubs, I created new asynchronous method which gets the clubs from the database and makes them to list.
GetClubCity with parameter city.
Again asynchronous method that checks in the clubs database where the address contains city and converts it to asynchronous list.
GetIdBy method with parameter Id.
In the clubs database search for  including the address, because is navigation property and selects firstOrDefault.
Save method I want to get info back that is greater than 0. The same methods are in the RaceRepository. After this manipulation, in the clubController I removed the database context and assign ClubRepository. Now I'm calling the methods in the club Repository. I made the Index method asynchronous and assign GetAllClubs method. I make the Detail method asynchronous too and assign GetIdByAsync method with parameter Id. The same stays for RaceController. Now everything seems better. Clean and readable code.