Add method - Added
New Method in the ClubController added Create. After that I create a View "Create" using the Bootstrap library. Binding data with asp-for ,  validate the form with the tag helper - asp-validation-for. On the form tag I added asp-action equals to "Create" and the method to post. In ClubController the method is assigned with attribute HTTPPOST & checks if the modelState is not valid. After adding the club to the list, it redirects to the index page.