Race Create method - Added
In RaceController, I Created "Create" View method. Then, I created asynchronous "Create" Method with parameter "race". Check if the ModelState is not valid and adding the race to the repository. At the end I'm redirected to the Index View. In RaceRepository, I had some issues, I have typed "return true" instead of "return Save()" to save the Races, I add or Delete. I Copied the same "Create" View of "Club" and started to replace the model using, and the properties that need to be replaced. After that is done, I tested it and it works.