Photo Upload - Cloudinary
Using cloudinary internet storage for images. Firstly in appsettings.json, I use
"CloudinarySettings": and set the CloudName,ApiKey & ApiSecret. IMPORTANT, if you want to work, you need to implement it, I delete my info for security reasons. Then, I create CloudinarySettings class and write three properties CloudName,ApiKey & ApiSecret. Secondly, I create interface "IPhotoService", that has two methods for AddingPhoto and DeletingPhoto. Then, I create Services folder and in there, I create PhotoService Class that inherit the interface "IPhotoService". Importing the options configuration in and create new account using the Cloudinary options. For AddPhoto method, I use built in function in Cloudinary to get the options I want.  At the end I UploadAsync() the photo. For the Delete method is easier, using the Destroy built in function. IMPORTRANT ,before doing all these thing, I Install the CLoudinary package from nuget packages. Then, I create ViewModels folder and add "CreateClubViewModel" for clubs and "CreateRaceViewModel" for races. The most specific thing in the ViewModels is that, the image is type of IFormFile. Now in the ClubController, I import the photoServices and now I'm using the CreateClubViewModels. In result variable i take the photoService with the built in function AddPhotoAsync() with paramer clubVM.Image. Then, I create new Club and add the club to the clubRepository. Validation check and return the clubVM. Same is for the RaceController. Now in the program.cs , I configure the CloudinarySettings, to use the CloudinarySettings in the appsettings.json file and import the interface photoServices. IMPORTANT in the "create" view of race and club, in the form tag, I added enctype="multipart/form-data". and make the input for the image type equal to "file". IMPORTANT! As I said to make it work you need to Implement your own, CloudName,ApiKey & ApiSecret.