Edit Method (Part 1)
I Created new ViewModel for Edit. In ClubController, I create Edit method with parameter (int id) In the method, I get the club by it's repository with method GetIdByAsync and pass the parameter inside. I check if the club is null, to get the current club, that I want to edit. Then, I create new ClubViewModel with the properties inside. At the end I return the ClubViewModel. Now, I created new View - Edit and copied the form of the "Create" View, with small adjustments like - creating one input after the form tag that is of type - hidden. and taking the AddressId as a value. Changed the asp-action to "Edit". Now in the "Index" view of club, I made the edit button interactive using the asp tag helpers. (Part 2) of the Edit method is comming.