Edit method (PART 2) - Ready
Creating HttpPost method Edit.
Firstly, checks if the model state is invalid. After that validation, I get the userClub from the repository asNoTracking!. Then If the userClub is not null , i go through "try & catch", firstly in the photoService, I "delete" the old photo and after that attaches the new photo from "cloudinary". Then the club is updated and returns to the Index View. In the IClubRepository, I copied the method GetIdByAsync, but this time using the AsNoTracking() funtion. I renamed the method - GetIdByAsyncNoTracking. I did this because, I got error - "another instance with the same key "id" is already being tracked" and to fix this I must make new method with NoTracking function.  In the clubEditViewModel, I make the URL optional.