Edit Method in RaceController - Done
Created new ViewModel, the only difference is that in Races, I use RaceCategory the same in the Edit View. Connect the edit button, using tag helpers. In the RaceController, two new methods one Edit, another one Edit which uses the Post request. Again in the Interface, I created new NoTracking Method, otherwise cannot be tracked, because another instance is tracked with the same key value id. Of course, I wrote it in the RaceRepositories. At the end of the post method in the RaceController, I update the database and everything that is changed,it's saved in the database.