Login Page
I created new class in the ViewModels - LoginViewModel, using validation annotations. Created AccountController, with 2 Login methods, one get, the other post. In constructor, I get the information about userManager, signInManager and the context. Firstly in the POST method there is a validation and after that, I get the user email. Then, I check if the user is not null and after that, I check if the password is the same as it should be to sign in. if one of this is not true, it returns view Error TempData with nice error text. New Login view added, bootstrap form and validation.