Register Page, Added
Almost the same code like in the Login Page, but with small key adjustments, I make the addressId nullable with ? operator, in the AppUser. In the RegisterController, I set the userName to the Email. Another method, I added is Logout, that signout of account and redirects to Index page of Race. I had to apply new migration to make the addressId nullable. In the Register View , I imported the using, that needs to fire the code and import form with validation. Now you can register and login.