Dashboard explanation
In the "ClubController" & "RaceController", I created delete method. New view "Delete" & added an anchor tag like button to be pressed. Then, I created new interface -IDashRepository with methods, to get all the clubs and races. After that, I create DashRepository and implement the two methods, using the database and HttpAccessor. Both methods are the same, i took the current user from the httpaccessor and I check if the AppUser-Id is equal to the current User. At the end, I return list of user clubs, or races. Then, I created new ViewModel - DashboardViewModel, including list of races and list of clubs as properties. In the program.cs, I registered the dashboard interface and start to use the authentication. In the layout file, I make if check - if the user is authenticated can see the logout button "anchor tag" and the dashboard. In "Race" index & "Club" Index, I check if the user is authenticated and it's in the role of admin. Only the admin can edit or delete the clubs and the races. I added new "dashboard" Index view. Now the second part of the dasboard is comming.