Dashboard (Part 2). Done
Firstly, I created my extension, GetUserId(), using the claimtypes "using". In "ClubController", I'm using the HttpAccessor. In "Create" method, I get the current id by the Accessor and the extension, that I created. Then, I create new "CreateClubViewModel", that appUserId = to the curUserId. Return the createClubVIewModel at the end. In the HttpPost, I added the AppUserId. Same in the "RaceController" In the ViewModels of Create, I added the AppUserId property. Important in "Create" View of Race and Club, I added the input of type hidden, that gets the AppUserId with tag helper. In the DashboardRepository instead of using the ToString method, I used my extension.