User details, update
Created new UserViewModel, with 4 properties. I make the index method asynchronous Task<IActionResult>, firstly in the method, I get the all users by the UserRepository. Then, I created list of UserViewModel and mapped every property. After that, I add the users to the list and at the end, I return the list. In the AppUser Model, I have written "Milage", instead of "Mileage", so, I replaced it and then, I add migration and update the database. In the program.cs, I registered the "UserRepository", now the View part left.