User detail page done
Created new ViewModel - UserDetailViewModel, with the same properties as the UserViewModel for now. In the UserController, I get the user by id and creating new ViewModel, at the end I return the VIewModel. In the "Index" method, I had an error, I must loop through the users ,not to the result. New View Detail, using bootstrap, I made the forms to look nice.