Edit User Profile
In "DashboardController", I bring the httpContextAccessor and the photoService. Then, I created new method EditUserProfile() and in that method, I take the id of the user and check if the user exists. At the end, I return the EditUserViewModel, that I created. In the Repository. I implement the interface method. Added three new properties in the "EditUserViewModel" and added migration,then updated it and now everything is working. Now I'm working on adding a form to edit the user profile