Edit user profile - done
Now the users can edit their own profile and add a picture to it. In "DashboardController", I created my own mapped, that uses the AppUser model, EditUserVIewModel and the ImageUploadResult. Mapped the properties using the "user" and the "editVM". Then , I created [HttpPost] method, with parameter EditViewModel. Some if checks in the method. Updating the user info and redirect to the Index action. New methods in the DashboardRepository GetIdByNoTracking, Update and more.