IpInfo.. Real city and state
Got the Ip Address from IPInfo.io
New ViewModel added for the Ip Configuration. Little improvements in the Index View of HomeController. In HomeController, In Try Catch I set the settings for right configuration for the Ip. One If check and at the end, I return the ViewModel. To make this work on you, you have to sign up in IpInfo.io to get your IpAddress.