GetPokemonRating UnitTest
Getting the id first, after that faking the list of pokemons dto's,  call the Pokemon Repository and call the method u want to test, pass the parameter that needs and pass in the return what is needed.  Then adding the controller. In result parameter, I call the controller methods with passed parameter inside. At the end , I want to return some value but not null.