Detail page
Detail page added, creating method Detail that has parameter int id.
Get the clubs and the most important here is that I'm using the include function of EntityFramework.Core to get the address for the clubs, the same method stays for the races. Then created new View Detail. Using bootstrap to make it look good and then using the model Club for clubs and model Race for races. Getting the real data with @symbol and item. Making the bootstrap buttons dynamic. Firstly I make the button an anchor tag, then I use asp properties to lead the functions.