Uploading photo .. done
NOTE - to use the application, you must have register in Cloudinary, to get access to you - CloudName, ApiKey, ApiSecret. In PhotoService, I make the Delete method for photos , with the cloudinary package function - Destroy. Created DTO'S , ViewModels folder and because of the VIewModels everything becomes much easier. In Views , implement the right ViewModels. In ClubController, I check if the modelstate is valid, and if it is , I added the club to the repository. Same stays for the Races. In Program.cs, I configure the cloudinary settings and register the IPhotoService.