Edit
Edit function, created. Now the clubs and the races can be edited. In the Repositories, I created one more method GetIdByNotTracking. Then , I created new ClubEditViewModel and new RaceEditViewModel, after that in the controllers , I created one HttpGet and another HttpPost method.  In the post method there is a validation part and one try catch, Firstly the photo is deleted when you are editing and you must add another photo, else you cannot proceed. That stays for races and clubs