Login Created
New ViewModel - Login, In AccountController, I registered the UserManager,SignInManager and the database. One Get method and Post .
The post method starts by checking the ModelState. Then the user variable gets finds in the database the user with the same email, and if it's not found,it gives an error, the same stays for the password, but it's not from the database. At the end if everything succeed it will return you the index page of the races. New Login View added with validation inside and beautiful style, using bootstrap.