Register & Logout added
New RegisterViewModel creating, 3 properties inside with data validation. In the layout, I added link to the register action. New View - Register, with form inside. In the AppUser class, the AddressId, i made it nullable. In AccountController, I check if the ema9l address is already used or not.  If it's used an error comes, else  the user is created, with role User and is redirected to the races.