Authentication & Dashboard
Authentication added in the views of index races and index clubs. In the layout page too, if the user is logged in, can see the dashboard and logout, but if it's not logged in, can register and the login in. In the program.cs file , I added use authentication, to activate the authenticating features like IsInRole - Admin or User.