User - List/Detail page, done.
You can see all the users in application, when you are in role admin, you can edit or remove all of them. New ViewModels, UserView & UserDetail. New Detail View and Index View Page of User updated. New interface of user, I'm not using all of the methods,  In the UserController, I get the userRepository, then using the method from the repository get all users, adding the viewmodel to the result and returning the list of users. Now the Detail , I'm getting the User by the specific Id and then returning all his properties.