Ip Location
Now the application, have the access to your location , state and city. Which means in the first page, when you are in the application, you can see running clubs near your city. For that purpose you need to have registration in ipinfo to get your current ip url. You need to paste it in HomeController, where the string url is equals to "". Small modifications in Index View file.