Edit method added.
Now using ViewModels.  Adjustment in the Create method, making it to work with ViewModels. For edit method, I'm using new method from the repository , the id to not be track. Basically the same method as getidbyasync, but without tracking.  In Edit.  using the ViewModels again, assign the values and the update the object., at the end I'm redirecting to index.