Login,Register and Logout
Created Login method, with the help of userManager and singInManager, I completed the method, three if checks, the first one if the user exists, the second if the user passwords doesn't match and the third if the user successfully logged in. In Register method, I check again if the user exists  and if exists, an error message comes. Then, I bind the Email and Username , to the EmailAddress of the RegisterViewModel, in the end, I create new user. The logout is the most simple one.  Changes in the views too, now became so much better application.