Adding dashboard
Created new DashboardViewModel, that have list of cars. Then, I created another one ViewModel, for editing the dashboard, that includes information about the user. Added new class that is static and have method GetUserId. This extension is needed for the dashboardRepository. In interfaces, I added another one for the dashboard, and after that, I created repository and inherited all the methods from the interface. In program.cs, I "registered" the IDashboardRepository and now the settings are almost ready. After that, I created DashboardController, that is using the httpContextAccessor and the dashboard repository. After that, I removed the key id in the owner class. Now in the controller, I added new method MapUserEdit and did the index. Finally, I created Index view of the dashboard.